WIP: Cleanup dependencies between cli and mcp server
