+++ .gitignore
@@ -131,4 +131,4 @@ dist
 .pnp.*
 
 # Turbo
-.turbo/cache
+**/.turbo
