+++ packages/cli/package.json
@@ -3,9 +3,8 @@
   "version": "3.1.6",
   "description": "CLI tools for responsible-vibe development workflows",
   "type": "module",
-  "main": "dist/index.js",
   "bin": {
-    "responsible-vibe-cli": "dist/visualization-launcher.js"
+    "responsible-vibe-cli": "dist/index.js"
   },
   "scripts": {
     "build": "tsc -p tsconfig.build.json",
