+++ packages/cli/src/config-generator.ts
