+++ packages/mcp-server/tsconfig.json
@@ -4,7 +4,8 @@
     "outDir": "dist",
     "rootDir": "src",
     "exactOptionalPropertyTypes": false,
-    "noPropertyAccessFromIndexSignature": false
+    "noPropertyAccessFromIndexSignature": false,
+    "noEmit": false
   },
   "include": ["src/**/*"],
   "exclude": ["dist", "node_modules"]
